+++ ristretto255/utils.h
@@ -25,7 +25,7 @@ void print_32(const u8* o);
 
 uint32_t bytes_eq_32( const u8 a[BYTES_ELEM_SIZE],  const u8 b[BYTES_ELEM_SIZE]);
 
-void crypto_wipe(void *secret, size_t size);
+void crypto_wipe(void *secret, int32_t size);
 
 
 // The macro WIPE_BUFFER is used throughout this file,
